sap_hana_install: Revert the hdblcm find logic; solve issue #39
For the time being, we support the SAP HANA server software only in
IMBD*SAR files, in which case the hdblcm command will be in directory
SAP_HANA_DATABASE. As it might occur that there are different
hdblcm files in different subdirectories, it appears to be more
appropriate to only search for directory SAP_HANA_DATABASE for
selecting the hdblcm command.

Relates to issue #29

This commit also solves issue #39: We want to show the complete
hdblcm command line before executing it. The command line also
contains the path name of the configfile so it is easier to
check it during execution.