sap_hana_install: Improve SAR files handling (#40)
If there is more than one SAR file in the software directory, moving the
contents of the second file of the same software to the correct location
fails and the role aborts.
So we introduce a new variable, `sap_hana_install_sarfiles_list`, which
can be used to provide a list of SAR files to extract.

Solves issue #40.