sap_general_prec: Improve RHEL repo handling
This commit further improves the RHEL repo handling:
The role now automatically uses the correct repo names and selects those
with the highest support life time (e.g. normal repos for RHEL 8.3 and
e4s repos for RHEL 8.4).

NetWeaver, HANA, and HA repos can be deselected from the list of repos
to be enabled, and it is possible to specify if all other repos should
be disabled before enabling the repos.

It is also possibe to specify a customized list of repos, which is
important for cloud systems.

This commit also includes the assert code.

As part of this effort, the main files for installation and configuration
have been moved from the main task directory to OS specific
subdirectories.

Relates to issues: #35 and #41