sap_general_preconfigure: Rewrite SELinux handling
The adds detecting the current SELinux status and set the reboot requirement
flag if necessary. So if the static configuration matches the desired
configuration and the current status is different, the reboot requirement
flag will be set as well.

Also the grubby command will now only be called if necessary.

Solves issue #47.