sap_hana_install: Rewrite checksum verification
With this commit, the checksums of SAPCAR EXE files are verified even
before the version detection. Also, the checksum verification now
supports a central checksum file.
Relates to: #67

Additional tags have been added to allow further limitation of the
activities performed by the role. This is now the list of supported
tags:
- sap_hana_install_hdblcm_commandline
- sap_hana_install_preinstall
- sap_hana_install_prepare_sapcar
- sap_hana_install_prepare_sarfiles
- sap_hana_install_chown_hana_directories (mainly to be used with
  --skip-tags when using tag sap_hana_install_preinstall)
- sap_hana_install_extract_sarfiles (mainly to be used with
  --skip-tags when using tag sap_hana_install_prepare_sarfiles)
Relates to: #58

Files created by the template module should have a note at the top
indicating that this file should not be manually modified.
Relates to: #66

Also all Ansible modules have been renamed to FQCN.
Relates to: #68