sap_install_media_detect: Use a one step file type determination method
The new method uses a tool named sapfile to examine the SAP and UNIX
file types and populate a dictionary. Each file is stored with
directory, file name, SAP file type, and archive file type.
Using this method, the SAP file types have to be determined only once.
The files are then copied to local storage (in case the software is
located on a remote NFS file system), extracted and moved properly
according to their entries in the dictionary.
As the last step, the variables for SWPM are set using the entries for
SAP file type in the dictionary, as these are identical to the
subdirectory names.