sap_install_media_detect: rename some variables
With identical variables in `find_files_after_extraction.yml`, some of
them can be overwritten as follows:
    "detect_directory_export_extracted": {
        "changed": false,
        "skip_reason": "Conditional result was False",
        "skipped": true
    }
These variables have to have different names so that the assignment of
final variables in `set_global_vars.yml` is done correctly.

This error did not happen with the previous version because each
detection was done in a separately included file.