sap_install_media_detect: Improve the handling of file renaming
The files without endings can now be renamed in the source directory
even in case of remote_dir, if the source directory is writable.
If the files without endings are located on a source directory which is
not writable, the role will fail and explain the reason for the failure.