sap_install_media_detect: Keep non-sapfile supported directories
With the output of 'sapfile -s', we are able to get the names of
directories into which the role moves SAP archive files. So the role can
now move files from there back to the top level and remove those
directories. The extracted directories are also (recursively) removed at
this stage.

The above is needed for idempotency (meaning that the role can run
twice or more and the result is always identical).

All other directories remain in place, including their contents.

Solves issue #523.