various roles: Fix issue sap-linuxlab#556 for non-assert runs
This commit makes the roles compatible with the following versions of
`ansible-core`:
- 2.16.1
- 2.15.8
- 2.14.12
when running in normal (=non-assert) mode.

The preconfigure roles also support an extended check mode, called
assert mode, in which the roles do not change anything but verify
if all settings are correct. This commit will not cover the assert
mode. So several tasks of the preconfigure roles in assert mode will
fail. By using the `_assert_ignore_errors` role parameters in assert mode,
the roles will not fail but the affected tasks will not be executed so
those settings will not be validated.

Relates to issues #555 and #556.

Signed-off-by: Bernd Finger <bfinger@redhat.com>