sap_swpm: Prioritize sap_swpm_templates_install_dictionary members
Due to the use of global variables in sap_swpm, there appears to be no
simple way of giving priority to top level vars over the same variables
defined in sap_swpm_templates_install_dictionary.
We can also assume that if using sap_swpm_templates_install_dictionary,
this is done on purpose and with the intention to use its members with
priority over top level variables.

Signed-off-by: Bernd Finger <bfinger@redhat.com>