sap_hana_install: Set the install execution mode to 'optimized'
This is required when installing the HANA component 'LSS', which is the
default in HANA 2.0 SPS08. The installation execution mode can be set to
the previous value of 'standard' by using the role variable
'sap_hana_install_install_execution_mode'. See the file
`defaults/main.yml` of this role.

Solves issue #894.

Signed-off-by: Bernd Finger <bfinger@redhat.com>