sap_ha_pacemaker_cluster: fix internal-error
For some reason the LSR collection variable is in the way of the ansible
syntax-check of some tools, when the task that calls the ha_cluster LSR
uses `import_role`.
The same syntax works without errors when the task uses `include_role`
instead.

Fixes galaxy-importer error "internal-error: Unexpected error code 1".