sap_ha_pacemaker_cluster: fix SAPStartSrv monitor
The SAPStartSrv monitor operation should not run on an interval, only
for a probe once at start. The added operation parameters will make sure
that the monitor operation is disabled.
To make it very clear that the interval=0 is not a mistake, the
`enabled` parameter is set to `false` as well.