collection: Cleanup the changelog(s)
In this commit, the file changelogs/changelog.yml is being reformatted
to comply with the standards as published in
https://github.com/ansible-community/antsibull-changelog/blob/main/docs/changelogs.md
and
https://github.com/ansible-community/antsibull-changelog/blob/main/docs/changelog.yaml-format.md
.

The command:

```
antsibull-changelog lint-changelog-yaml changelogs/changelog.yaml
```

returns with return code 0.

The command:

```
antsibull-changelog generate
```

generates the file CHANGELOG.rst from file changelogs/changelog.yaml.
The file CHANGELOG.rst is also part of this commit.

Signed-off-by: Bernd Finger <bfinger@redhat.com>