sap_hana_preconfigure: Prioritize old selinux and faplicyd vars
If sap_hana_install_modify_selinux_labels is defined in a playbook or inventory,
use that one instead of sap_hana_install_configure_selinux (which is also defined
in defaults/main.ylm).

If sap_hana_install_use_fapolicyd is defined in a playbook or inventory,
use that one instead of sap_hana_install_configure_fapolicyd (which is also defined
in defaults/main.ylm).

As a consequence, when using the new variables, the old ones should be
removed from the playbook or inventory - otherwise, the new variables
will be ignored.

Relates to #1134.

Signed-off-by: Bernd Finger <bfinger@redhat.com>